+++ checks/check_voronoi.py
@@ -1,4 +1,7 @@
 from __future__ import print_function, division
+
+import numpy as np
+
 import imgaug as ia
 import imgaug.augmenters as iaa
 
@@ -10,16 +13,19 @@ def main():
         iaa.RegularGridPointsSampler(n_rows=50, n_cols=50),
         0.5
     )
-
     uniform_sampler = iaa.UniformPointsSampler(50*50)
 
-    aug = iaa.Voronoi(point_sampler=reggrid_sampler, p_replace=1.0,
-                      max_size=128)
-    ia.imshow(aug(image=image))
+    augs = [
+        iaa.Voronoi(points_sampler=reggrid_sampler, p_replace=1.0,
+                    max_size=128),
+        iaa.Voronoi(points_sampler=uniform_sampler, p_replace=1.0,
+                    max_size=128),
+        iaa.UniformVoronoi(50*50, p_replace=1.0, max_size=128)
+    ]
 
-    aug = iaa.Voronoi(point_sampler=uniform_sampler, p_replace=1.0,
-                      max_size=128)
-    ia.imshow(aug(image=image))
+    images = [aug(image=image) for aug in augs]
+
+    ia.imshow(np.hstack(images))
 
 
 if __name__ == "__main__":